# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: C4 Admin API
  version: 1.0.0
  description: '管理端 API（模板/账号/分组/规则/日志/统计）。认证：Authorization: Bearer <admin_token>。'
servers:
  - url: /api/admin
...
paths:
...
  /redemption-codes:
    post:
      operationId: PostRedemptionCodes
      tags: [redemptions]
      summary: 生成兑换码（1..count 个；count 默认 1，上限 1000；type 枚举/value>0/temp_balance 必填 resource_expires_at）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateRequest' }
      responses:
        '200':
          description: 生成的完整码列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerateResponse' }
        default:
          $ref: '#/components/responses/Error'
    get:
      operationId: GetRedemptionCodes
      tags: [redemptions]
      summary: 兑换码列表（增强分页范式 page/page_size；type/status 筛选；sort 白名单）
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1, description: 页码（1-based） } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 1000 } }
        - { name: type, in: query, schema: { type: string, enum: [balance, concurrency, temp_balance] } }
        - { name: status, in: query, schema: { type: string, enum: [active, disabled] } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: 兑换码列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RedemptionCodeListResponse' }
        default:
          $ref: '#/components/responses/Error'
  /redemption-codes/batch-deactivate:
    post:
      operationId: PostRedemptionCodesBatchDeactivate
      tags: [redemptions]
      summary: 批量失效兑换码（单事务；已失效 no-op；缺失 id → 404 含缺失详情）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchDeactivateRequest' }
      responses:
        '200':
          description: 新失效数（已 disabled 不计）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchDeactivateResponse' }
        default:
          $ref: '#/components/responses/Error'
  /redemption-codes/{id}/uses:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetRedemptionCodesIdUses
      tags: [redemptions]
      summary: 某码的兑换记录（审计；码缺失 → 404）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: 兑换记录列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RedemptionUseListResponse' }
        default:
          $ref: '#/components/responses/Error'
  /redemption-codes/{id}/deactivate:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      operationId: PostRedemptionCodesIdDeactivate
      tags: [redemptions]
      summary: 单码失效（已失效 no-op；缺失 → 404）
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeactivateRequest' }
      responses:
        '200':
          description: 失效成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeactivateResponse' }
        default:
          $ref: '#/components/responses/Error'
...
  /api/user/auth/register:
    post:
      operationId: PostUserAuthRegister
      tags: [user]
      summary: 注册（signup_enabled 开关检查；注册即登录返回 JWT）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserAuthRegister' }
      responses:
        '200':
          description: JWT + 用户
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserAuthResponse' }
        default:
          $ref: '#/components/responses/Error'
  /api/user/auth/login:
    post:
      operationId: PostUserAuthLogin
      tags: [user]
      summary: 登录（bcrypt 校验 → JWT）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserAuthLogin' }
      responses:
        '200':
          description: JWT + 用户
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserAuthResponse' }
        default:
          $ref: '#/components/responses/Error'
...
  /api/user/auth/register-code:
    post:
      operationId: PostUserAuthRegisterCode
      tags: [user]
      summary: 发送注册验证码（signup_enabled  gate 403；已注册静默抑制仍 200；限频 429）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterCodeRequest' }
      responses:
        '200':
          description: 已发送（恒 {sent:true}）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SentResponse' }
        default:
          $ref: '#/components/responses/Error'
  /api/user/auth/forgot-password:
    post:
      operationId: PostUserAuthForgotPassword
      tags: [user]
      summary: 忘记密码发码（恒 200 {sent:true}，反枚举——无论账号是否存在或邮件是否启用）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '200':
          description: 已发送（恒 {sent:true}）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SentResponse' }
        default:
          $ref: '#/components/responses/Error'
  /api/user/auth/reset-password:
    post:
      operationId: PostUserAuthResetPassword
      tags: [user]
      summary: 重置密码（验证码校验→更新密码；成功后立即撤销该用户既有 JWT）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200':
          description: 重置成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChangePasswordResponse' }
        default:
          $ref: '#/components/responses/Error'
  # --- /user 面业务（tag: user；JWT 保护——Router 中间件，仅本人数据） ---
  # 注意 chi 路径顺序：/user/keys 必须先于 /user/keys/{id} 注册（生成器按 spec 顺序）。
...
components:
  responses:
    Error:
      description: 错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    UserRole:
      type: string
      enum: [platform_admin, user]
    UserStatus:
      type: string
      enum: [active, disabled]
    User:
      type: object
      properties:
        ID: { type: integer, format: int64 }
        Email: { type: string }
        Role: { $ref: '#/components/schemas/UserRole' }
        Status: { $ref: '#/components/schemas/UserStatus' }
        MaxConcurrency: { type: integer }
        Balance: { type: number, format: double, description: "余额 USD（浮点；内部存储毫分——1 USD = 100,000 毫分，API 边界换算）" }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
    UserAuthRegister:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
        code: { type: string, description: "可选：邮箱验证码（mail.register_verification 开启时必填，否则 400 email verification required）" }
    UserAuthLogin:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
...
    RegisterCodeRequest:
      type: object
      required: [email]
      properties:
        email: { type: string }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string }
    ResetPasswordRequest:
      type: object
      required: [email, code, new_password]
      properties:
        email: { type: string }
        code: { type: string }
        new_password: { type: string }
    SentResponse:
      type: object
      required: [sent]
      properties:
        sent: { type: boolean }
...
    ChangePasswordResponse:
      type: object
      required: [updated]
      properties:
        updated: { type: boolean }
    UserAuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/User' }
...
    RedemptionType:
      type: string
      enum: [balance, concurrency, temp_balance]
    RedemptionStatus:
      type: string
      enum: [active, disabled]
    RedemptionCode:
      type: object
      required: [ID, Code, Type, Value, MaxUses, UsedCount, Status, CreatedBy, CreatedAt, UpdatedAt]
      properties:
        ID: { type: integer, format: int64 }
        Code: { type: string, description: XXXX-XXXX-XXXX-XXXX（16 字符，生成后不可编辑，仅可失效） }
        Type: { $ref: '#/components/schemas/RedemptionType' }
        Value: { type: number, format: double, description: 面值：balance/temp_balance = USD（1 = $1）；concurrency = 并发数（整数）。存储毫分（API 边界换算） }
        Remark: { type: string, nullable: true }
        ExpiresAt: { type: string, format: date-time, nullable: true, description: 码未兑换即过期；null = 永久 }
        ResourceExpiresAt: { type: string, format: date-time, nullable: true, description: 兑换后资源到期；temp_balance 必有，其余恒 null }
        MaxUses: { type: integer, description: 1 = 单次码；>1 = 多人码 }
        UsedCount: { type: integer, description: 已兑换次数 }
        Status: { $ref: '#/components/schemas/RedemptionStatus' }
        CreatedBy: { type: integer, format: int64, description: 0 = 系统（静态 admin token）；>0 = platform_admin 用户 id }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
    RedemptionUse:
      type: object
      required: [ID, CodeID, UserID, Value, CreatedAt]
      properties:
        ID: { type: integer, format: int64 }
        CodeID: { type: integer, format: int64 }
        UserID: { type: integer, format: int64 }
        Value: { type: number, format: double, description: 兑换时的值快照（同面值单位语义） }
        ResourceExpiresAt: { type: string, format: date-time, nullable: true }
        CreatedAt: { type: string, format: date-time }
...
    RedemptionCodeListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/RedemptionCode' }
    RedemptionUseListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/RedemptionUse' }
...
    GenerateRequest:
      type: object
      required: [type, value]
      properties:
        type: { $ref: '#/components/schemas/RedemptionType' }
        value: { type: number, format: double, description: 面值：balance/temp_balance = USD；concurrency = 并发数（整数）。> 0 }
        remark: { type: string }
        expires_at: { type: string, format: date-time, description: 码未兑换即过期；缺省 = 永久 }
        resource_expires_at: { type: string, format: date-time, description: 兑换后资源到期；temp_balance 必填（决策 4） }
        max_uses: { type: integer, description: 1 = 单次码（缺省）；>1 = 多人码 }
        count: { type: integer, description: 生成个数 1-1000（缺省 1） }
    GenerateResponse:
      type: object
      required: [codes]
      properties:
        codes:
          type: array
          items: { $ref: '#/components/schemas/RedemptionCode' }
    DeactivateRequest:
      type: object
      description: 单码失效无请求体（保留空 schema 对齐批量端点命名）
    DeactivateResponse:
      type: object
      required: [deactivated]
      properties:
        deactivated: { type: boolean }
    BatchDeactivateRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items: { type: integer, format: int64 }
          minItems: 1
          maxItems: 100
    BatchDeactivateResponse:
      type: object
      required: [deactivated]
      properties:
        deactivated: { type: integer, description: 新失效数（已 disabled no-op 不计） }
    # --- litellm 模型价格（Phase 5 计费价格来源） ---
...